feat: implement InfluenceDiagramService and related interfaces for managing influence diagrams
