+++ PrismaDotnetApi/PrismaApi.Application/Services/ProjectService.cs
@@ -14,37 +14,22 @@ public class ProjectService : IProjectService
 {
     private readonly IProjectRepository _projectRepository;
     private readonly IProjectRoleRepository _projectRoleRepository;
-    private readonly IIssueRepository _issueRepository;
-    private readonly IEdgeRepository _edgeRepository;
     private readonly IUserRepository _userRepository;
     private readonly IBoardSheetRepository _boardSheetRepository;
-    private readonly IDiscreteProbabilityRepository _discreteProbabilityRepository;
-    private readonly IDiscreteUtilityRepository _discreteUtilityRepository;
-    private readonly IRestrictionTableRepository _restrictionTableRepository;
     private readonly IMemoryCache _cache;
 
     public ProjectService(
         IProjectRepository projectRepository,
         IProjectRoleRepository projectRoleRepository,
-        IIssueRepository issueRepository,
-        IEdgeRepository edgeRepository,
         IUserRepository userRepository,
         IBoardSheetRepository boardSheetRepository,
-        IDiscreteProbabilityRepository discreteProbabilityRepository,
-        IDiscreteUtilityRepository discreteUtilityRepository,
-        IRestrictionTableRepository restrictionTableRepository,
         IMemoryCache cache)
     {
         _projectRepository = projectRepository;
         _projectRoleRepository = projectRoleRepository;
-        _issueRepository = issueRepository;
-        _edgeRepository = edgeRepository;
         _userRepository = userRepository;
         _cache = cache;
         _boardSheetRepository = boardSheetRepository;
-        _discreteProbabilityRepository = discreteProbabilityRepository;
-        _discreteUtilityRepository = discreteUtilityRepository;
-        _restrictionTableRepository = restrictionTableRepository;
     }
 
     public async Task<List<ProjectOutgoingDto>> CreateAsync(List<ProjectCreateDto> dtos, bool createDefaultRole, UserOutgoingDto userDto, CancellationToken ct = default)
@@ -124,33 +109,6 @@ public class ProjectService : IProjectService
         return projects.ToPopulatedDtos();
     }
 
-    public async Task<InfluenceDiagramDto> GetInfluenceDiagramAsync(Guid projectId, UserOutgoingDto user, CancellationToken ct = default)
-    {
-        var cachedDiagram = _cache.GetCacheItemAsInfluenceDiagram(projectId, user);
-        if (cachedDiagram != null)
-        {
-            return cachedDiagram;
-        }
-
-        var issueEntities = await _issueRepository.GetIssuesInInfluenceDiagram(projectId, IssuesUserFilter(user), ct);
-        var edgeEntities = await _edgeRepository.GetEdgesInInfluenceDiagram(projectId, EdgesUserFilter(user), ct);
-        var discreteProbabilities = await _discreteProbabilityRepository.GetAllAsync(filterPredicate: e => e.Uncertainty!.Issue!.ProjectId == projectId, ct: ct);
-        var discreteUtilities = await _discreteUtilityRepository.GetAllAsync(filterPredicate: e => e.Utility!.Issue!.ProjectId == projectId, ct: ct);
-        var restrictionTables = await _restrictionTableRepository.GetAllAsync(filterPredicate: e => e.ProjectId == projectId, ct: ct);
-        var diagram = new InfluenceDiagramDto
-        {
-            projectId = projectId,
-            issues = issueEntities.ToOutgoingDtos(),
-            edges = edgeEntities.ToOutgoingDtos(),
-            discreteProbabilities = discreteProbabilities.ToDtos(),
-            discreteUtilities = discreteUtilities.ToDtos(),
-            restrictionTables = restrictionTables.ToOutgoingDtos(),
-        };
-
-        _cache.AddCacheItem(new CacheItem { CacheKey = CacheKeys.GetInfluenceDiagramKey(projectId) }, CacheConstants.DefaultMediumQueryCacheInTimeSpan, diagram);
-        return diagram;
-    }
-
     private async Task EnsureDefaultSheetsExists(List<Guid> projectIds, UserOutgoingDto user, CancellationToken ct = default)
     {
         var existingSheets = await _boardSheetRepository.GetAllAsync(withTracking: false, filterPredicate: e => projectIds.Contains(e.ProjectId), ct: ct);
@@ -171,12 +129,6 @@ public class ProjectService : IProjectService
 
     private static Expression<Func<Project, bool>> UserFilter(UserOutgoingDto user)
         => e => e.Public || e.ProjectRoles.Any(p => p.UserId == user.Id);
-
-    private static Expression<Func<Issue, bool>> IssuesUserFilter(UserOutgoingDto user)
-        => e => e.Project!.Public || e.Project!.ProjectRoles.Any(p => p.UserId == user.Id);
-
-    private static Expression<Func<Edge, bool>> EdgesUserFilter(UserOutgoingDto user)
-        => e => (e.HeadNode!.Issue!.Project!.Public || e.HeadNode!.Issue!.Project!.ProjectRoles.Any(p => p.UserId == user.Id)) && (e.TailNode!.Issue!.Project!.Public || e.TailNode!.Issue!.Project!.ProjectRoles.Any(p => p.UserId == user.Id));
 
     private void RegisterPublicProjectsInCache(List<ProjectOutgoingDto> projects)
     {
