+++ PrismaDotnetApi/PrismaApi.Domain/Dtos/InfluenceDiagramDto.cs
@@ -1,7 +1,4 @@
-﻿using System.Text.Json;
-using PrismaApi.Domain.Constants;
-
-namespace PrismaApi.Domain.Dtos;
+﻿namespace PrismaApi.Domain.Dtos;
 
 public class InfluenceDiagramDto
 {
@@ -11,155 +8,4 @@ public class InfluenceDiagramDto
     public required ICollection<DiscreteProbabilityDto> discreteProbabilities { get; init; }
     public required ICollection<DiscreteUtilityDto> discreteUtilities { get; init; }
     public required ICollection<RestrictionTableOutgoingDto> restrictionTables { get; init; }
-}
-
-public static class InfluenceDiagramDtoExtensions
-{
-    public static InfluenceDiagramDto DeepClone<InfluenceDiagramDto>(this InfluenceDiagramDto source)
-    {
-        var json = JsonSerializer.Serialize(source);
-        return JsonSerializer.Deserialize<InfluenceDiagramDto>(json)!;
-    }
-
-    public static List<DiscreteUtilityDto> GetRestrictedDiscreteUtilities(this InfluenceDiagramDto influanceDiagramDto, Guid RestrictionTableId)
-    {
-        return influanceDiagramDto.discreteUtilities.Where(du => du.UtilityId == RestrictionTableId).ToList();
-    }
-
-    public static void CreateRestrictedDiscreteUtilities(this InfluenceDiagramDto influenceDiagramDto, Guid RestrictionTableId, EdgeOutgoingDto edge)
-    {
-        var restrictionTable = influenceDiagramDto.restrictionTables.FirstOrDefault(rt => rt.Id == RestrictionTableId);
-        if (restrictionTable is null) return;
-
-        // create the utility issue and nodes and edges
-        var utilityNode = new NodeViaIssueOutgoingDto
-        {
-            Id = RestrictionTableId,
-            IssueId = RestrictionTableId,
-            Name = restrictionTable.Name,
-        };
-        var utilityIssue = new IssueOutgoingDto
-        {
-            Id = RestrictionTableId,
-            Name = restrictionTable.Name,
-            Type = IssueType.Utility.ToString(),
-            Utility = new UtilityOutgoingDto
-            {
-                ProjectId = influenceDiagramDto.projectId,
-                Id = RestrictionTableId,
-            },
-            Node = utilityNode
-        };
-        var nodeDto = new NodeOutgoingDto
-        {
-            Id = RestrictionTableId,
-            IssueId = RestrictionTableId,
-            Name = restrictionTable.Name,
-            Issue = new IssueViaNodeOutgoingDto
-            {
-                Id = RestrictionTableId,
-                Name = restrictionTable.Name,
-                Type = IssueType.Utility.ToString(),
-                Utility = new UtilityOutgoingDto
-                {
-                    ProjectId = influenceDiagramDto.projectId,
-                    Id = RestrictionTableId,
-                },
-            }
-        };
-
-        var edge1 = new EdgeOutgoingDto
-        {
-            Id = Guid.NewGuid(),
-            TailIssueId = influenceDiagramDto.issues.First(i => i.Node.Id == edge.TailId).Id,
-            TailId = edge.TailId,
-            TailNode = edge.TailNode,
-            HeadIssueId = utilityNode.IssueId,
-            HeadId = utilityNode.Id,
-            HeadNode = nodeDto
-        };
-        var edge2 = new EdgeOutgoingDto
-        {
-            Id = Guid.NewGuid(),
-            TailIssueId = influenceDiagramDto.issues.First(i => i.Node.Id == edge.HeadId).Id,
-            TailId = edge.HeadId,
-            TailNode = edge.HeadNode,
-            HeadIssueId = utilityNode.IssueId,
-            HeadId = utilityNode.Id,
-            HeadNode = nodeDto
-        };
-        influenceDiagramDto.issues.Add(utilityIssue);
-        influenceDiagramDto.edges.Add(edge1);
-        influenceDiagramDto.edges.Add(edge2);
-        foreach (var entry in restrictionTable.RestrictionEntries)
-        {
-            if (!entry.IsChildUncertainty && entry.ParentStateId is not null && entry.ChildStateId is not null)
-            {
-                // The child of the entry is an option, meaning it is a parent of the discrete utility, but the parent can be either an option or an outcome, 
-                // Check which one it is and add it to the appropriate list of parent ids
-                var parentOutcomeIds = entry.IsParentUncertainty ? new List<Guid> { (Guid)entry.ParentStateId } : new List<Guid>();
-                var parentOptionIds = entry.IsParentUncertainty ? new List<Guid>{(Guid)entry.ChildStateId} : new List<Guid> { (Guid)entry.ChildStateId, (Guid)entry.ParentStateId };
-                // create a discrete utility for the restricted option given the parent state
-                var discreteUtility = new DiscreteUtilityDto
-                {
-                    ProjectId = influenceDiagramDto.projectId,
-                    UtilityId = RestrictionTableId,
-                    ParentOptionIds = parentOptionIds,
-                    ParentOutcomeIds = parentOutcomeIds,
-                    UtilityValue = entry.RestrictionValue == 0 ? double.MinValue : 0
-                };
-                influenceDiagramDto.discreteUtilities.Add(discreteUtility);
-            }
-        }
-    }
-    
-    public static void ApplyRestrictions(this InfluenceDiagramDto influenceDiagramDto)
-    {
-        var restrictionEntriesUncertainties = influenceDiagramDto.restrictionTables.SelectMany(rt => rt.RestrictionEntries).Where(re => re.IsChildUncertainty).ToList();
-        var restrictionTablesDecisions = influenceDiagramDto.restrictionTables.Where(rt => !rt.RestrictionEntries.All(re => re.IsChildUncertainty)).ToList();
-        var discreteProbabilities = influenceDiagramDto.discreteProbabilities;
-
-        foreach (var table in restrictionTablesDecisions)
-        {
-            // skip if all entries have a restriction value of 1, meaning no restrictions
-            if (table.RestrictionEntries.All(re => re.RestrictionValue == 1)) continue;
-            var edge = influenceDiagramDto.edges.First(e => e.Id == table.EdgeId);
-            influenceDiagramDto.CreateRestrictedDiscreteUtilities(table.Id, edge);
-        }
-        foreach (var entry in restrictionEntriesUncertainties)
-        {
-            if (entry.IsChildUncertainty)
-            {
-                // child is an outcome, set the discrete probabilities that have that outcome/option as a parent to 0
-                var affectedProbabilities = discreteProbabilities.Where(
-                    dp => dp.OutcomeId == entry.ChildStateId && 
-                    entry.ParentStateId is not null && 
-                    (dp.ParentOptionIds.Contains((Guid)entry.ParentStateId) || dp.ParentOutcomeIds.Contains((Guid)entry.ParentStateId))).ToList();
-                foreach (var probability in affectedProbabilities)
-                {
-                    probability.Probability = probability.Probability * entry.RestrictionValue;
-                    // need to normalize the probabilities for the parent state after setting some to 0
-                    // solver hanldes the case where all probabilities for a parent state are set to 0
-                }
-            }
-        }
-        var probabilityRows = discreteProbabilities.GroupBy(dp =>
-            string.Join(",", dp.ParentOptionIds.OrderBy(x => x)
-                .Concat(dp.ParentOutcomeIds.OrderBy(x => x)))).ToList();
-        foreach (var row in probabilityRows)
-        {
-            int precision = 2;
-            var totalProbability = row.Sum(x => x.Probability);
-            if (totalProbability is null || Math.Round(totalProbability.Value, precision) == 0 || Math.Round(totalProbability.Value, precision) == 1) continue; // no need to normalize 
-            
-            // normalize the probabilities for this row, but leave out any probabilities that are already 0
-            // since they are restricted and should not be normalized
-            var nonZeroProbabilities = row.Where(x => x.Probability > 0).ToList();
-            var nonZeroTotalProbability = nonZeroProbabilities.Sum(x => x.Probability);
-            foreach (var probability in nonZeroProbabilities)
-            {
-                probability.Probability = probability.Probability / nonZeroTotalProbability;
-            }
-        }
-    }
 }
